Refactor Spec.Conway into  Spec.Core
